Updated Some styles and fonts
